server/requirements.txt
fastapi
uvicorn
python-multipart
requests
transformers
torch
Pillow
sentencepiece
pydicom
numpy
grad-cam
python-jose[cryptography]
passlib
argon2-cffi
bcrypt==4.0.1
cryptography
python-dotenv
opencv-python
python-swiftclient
protobuf
huggingface_hub
